refactoring: put installed stuff into a dedicated types package
