+++ src/Stack/Package.hs
@@ -93,14 +93,16 @@ import           Stack.Types.Dependency
                    , libraryDepFromVersionRange
                    )
 import           Stack.Types.EnvConfig ( HasEnvConfig )
+
+import           Stack.Types.Installed
+                   ( Installed (..), InstallMap, InstalledMap
+                   , installedToPackageIdOpt )
 import           Stack.Types.NamedComponent
                    ( NamedComponent (..), subLibComponents )
 import           Stack.Types.Package
-                   ( BioInput(..), BuildInfoOpts (..), InstallMap
-                   , Installed (..), InstalledMap, Package (..)
+                   ( BioInput(..), BuildInfoOpts (..), Package (..)
                    , PackageConfig (..), PackageException (..)
-                   , dotCabalCFilePath, installedToPackageIdOpt
-                   , packageIdentifier
+                   , dotCabalCFilePath, packageIdentifier
                    )
 import           Stack.Types.PackageFile
                    ( DotCabalPath, PackageComponentFile (..) )
@@ -293,7 +295,7 @@ generateBuildInfoOpts BioInput {..} =
   deps =
     concat
       [ case M.lookup name biInstalledMap of
-          Just (_, Stack.Types.Package.Library _ident installedInfo) ->
+          Just (_, Stack.Types.Installed.Library _ident installedInfo) ->
             installedToPackageIdOpt installedInfo
           _ -> ["-package=" <> packageNameString name <>
             maybe "" -- This empty case applies to e.g. base.
