+++ src/Stack/Types/Package.hs
@@ -8,11 +8,9 @@ module Stack.Types.Package
   , ExeName (..)
   , FileCacheInfo (..)
   , InstallLocation (..)
-  , InstallMap
   , Installed (..)
   , InstalledLibraryInfo (..)
   , InstalledPackageLocation (..)
-  , InstalledMap
   , LocalPackage (..)
   , MemoizedWith (..)
   , Package (..)
@@ -27,12 +25,7 @@ module Stack.Types.Package
   , dotCabalMainPath
   , dotCabalModule
   , dotCabalModulePath
-  , installedGhcPkgId
-  , installedLibraryInfoFromGhcPkgId
   , installedMapGhcPkgId
-  , installedPackageIdentifier
-  , installedToPackageIdOpt
-  , installedVersion
   , lpFiles
   , lpFilesForComponents
   , memoizeRefWith
@@ -71,13 +64,19 @@ import           Stack.Types.Component
 import           Stack.Types.ComponentUtils (toCabalName)
 import           Stack.Types.Dependency ( DepValue )
 import           Stack.Types.EnvConfig ( EnvConfig, HasEnvConfig (..) )
-import           Stack.Types.GhcPkgId ( GhcPkgId, ghcPkgIdString )
+import           Stack.Types.GhcPkgId ( GhcPkgId )
 import           Stack.Types.NamedComponent ( NamedComponent )
 import           Stack.Types.PackageFile
                    ( DotCabalDescriptor (..), DotCabalPath (..)
                    , StackPackageFile
                    )
 import           Stack.Types.SourceMap ( CommonPackage, FromSnapshot )
+import Stack.Types.Installed
+                   ( Installed(..), InstalledLibraryInfo(..), InstalledMap
+                   , InstallMap, PackageDbVariety(..), PackageDatabase(..)
+                   , InstalledPackageLocation(..), InstallLocation(..)
+                   , toPackageDbVariety, simpleInstalledLib
+                   )
 
 -- | Type representing exceptions thrown by functions exported by the
 -- "Stack.Package" module.
@@ -210,11 +209,6 @@ packageIdentifier p = PackageIdentifier (packageName p) (packageVersion p)
 
 packageDefinedFlags :: Package -> Set FlagName
 packageDefinedFlags = M.keysSet . packageDefaultFlags
-
--- | Type synonym representing dictionaries of package names for a project's
--- packages and dependencies, and pairs of their relevant database (write-only
--- or mutable) and package versions.
-type InstallMap = Map PackageName (InstallLocation, Version)
 
 -- | GHC options based on cabal information and ghc-options.
 data BuildInfoOpts = BuildInfoOpts
@@ -355,63 +349,6 @@ lpFilesForComponents components lp = runMemoizedWith $ do
   componentFiles <- lpComponentFiles lp
   pure $ mconcat (M.elems (M.restrictKeys componentFiles components))
 
--- | Type representing user package databases that packages can be installed
--- into.
-data InstallLocation
-  = Snap
-    -- ^ The write-only package database, formerly known as the snapshot
-    -- database.
-  | Local
-    -- ^ The mutable package database, formerly known as the local database.
-  deriving (Eq, Show)
-
-instance Semigroup InstallLocation where
-  Local <> _ = Local
-  _ <> Local = Local
-  Snap <> Snap = Snap
-
-instance Monoid InstallLocation where
-  mempty = Snap
-  mappend = (<>)
-
--- | Type representing user (non-global) package databases that can provide
--- installed packages.
-data InstalledPackageLocation
-  = InstalledTo InstallLocation
-    -- ^ A package database that a package can be installed into.
-  | ExtraPkgDb
-    -- ^ An \'extra\' package database, specified by @extra-package-dbs@.
-  deriving (Eq, Show)
-
--- | Type representing package databases that can provide installed packages.
-data PackageDatabase
-  = GlobalPkgDb
-    -- ^ GHC's global package database.
-  | UserPkgDb InstalledPackageLocation (Path Abs Dir)
-    -- ^ A user package database.
-  deriving (Eq, Show)
-
--- | Type representing varieties of package databases that can provide
--- installed packages.
-data PackageDbVariety
-  = GlobalDb
-    -- ^ GHC's global package database.
-  | ExtraDb
-    -- ^ An \'extra\' package database, specified by @extra-package-dbs@.
-  | WriteOnlyDb
-    -- ^ The write-only package database, for immutable packages.
-  | MutableDb
-    -- ^ The mutable package database.
-  deriving (Eq, Show)
-
--- | A function to yield the variety of package database for a given
--- package database that can provide installed packages.
-toPackageDbVariety :: PackageDatabase -> PackageDbVariety
-toPackageDbVariety GlobalPkgDb = GlobalDb
-toPackageDbVariety (UserPkgDb ExtraPkgDb _) = ExtraDb
-toPackageDbVariety (UserPkgDb (InstalledTo Snap) _) = WriteOnlyDb
-toPackageDbVariety (UserPkgDb (InstalledTo Local) _) = MutableDb
-
 newtype FileCacheInfo = FileCacheInfo
   { fciHash :: SHA256
   }
@@ -462,60 +399,6 @@ dotCabalGetPath dcp =
     DotCabalMainPath fp -> fp
     DotCabalFilePath fp -> fp
     DotCabalCFilePath fp -> fp
-
--- | Type synonym representing dictionaries of package names, and a pair of in
--- which package database the package is installed (write-only or mutable) and
--- information about what is installed.
-type InstalledMap = Map PackageName (InstallLocation, Installed)
-
-data InstalledLibraryInfo = InstalledLibraryInfo
-  { iliId :: GhcPkgId
-  , iliLicense :: Maybe (Either SPDX.License License)
-  , iliSublib :: Map StackUnqualCompName GhcPkgId
-  }
-  deriving (Eq, Show)
-
-installedLibraryInfoFromGhcPkgId :: GhcPkgId -> InstalledLibraryInfo
-installedLibraryInfoFromGhcPkgId ghcPkgId =
-  InstalledLibraryInfo ghcPkgId Nothing mempty
-
-simpleInstalledLib ::
-     PackageIdentifier
-  -> GhcPkgId
-  -> Map StackUnqualCompName GhcPkgId
-  -> Installed
-simpleInstalledLib pkgIdentifier ghcPkgId =
-  Library pkgIdentifier . InstalledLibraryInfo ghcPkgId Nothing
-
-installedToPackageIdOpt :: InstalledLibraryInfo -> [String]
-installedToPackageIdOpt libInfo =
-  M.foldr' (iterator (++)) (pure $ toStr (iliId libInfo)) (iliSublib libInfo)
- where
-  toStr ghcPkgId = "-package-id=" <> ghcPkgIdString ghcPkgId
-  iterator op ghcPkgId acc = pure (toStr ghcPkgId) `op` acc
-
--- | Type representing information about what is installed.
-data Installed
-  = Library PackageIdentifier InstalledLibraryInfo
-    -- ^ A library, including its installed package id and, optionally, its
-    -- license.
-  | Executable PackageIdentifier
-    -- ^ An executable.
-  deriving (Eq, Show)
-
-installedPackageIdentifier :: Installed -> PackageIdentifier
-installedPackageIdentifier (Library pid _) = pid
-installedPackageIdentifier (Executable pid) = pid
-
-installedGhcPkgId :: Installed -> Maybe GhcPkgId
-installedGhcPkgId (Library _ libInfo) = Just $ iliId libInfo
-installedGhcPkgId (Executable _) = Nothing
-
--- | Get the installed Version.
-installedVersion :: Installed -> Version
-installedVersion i =
-  let PackageIdentifier _ version = installedPackageIdentifier i
-  in  version
 
 -- | Gathers all the GhcPkgId provided by a library into a map
 installedMapGhcPkgId ::
